Some improvements
I change the byte list in the Huffman tree class to a bit array (see Utils.cs), because its much easier to work with that.
I also change the data type of the text and the encoded text from string to byte, so now it works with every UTF-8 character.